Pokus/requirements.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
requests>=2.32,<3.0
pydantic>=2.11,<3.0
python-dotenv>=1.1,<2.0
PyYAML>=6.0,<7.0
pytest>=8.4,<9.0
